+++ fiddles/angular2/fiddle-0020-LayoutContainer/src/header/AppHeader.js
@@ -1,0 +1,1 @@
+//# sourceMappingURL=AppHeader.js.map