+++ fiddles/angular2/fiddle-0020-LayoutContainer/src/navigation/Navigation.js
@@ -1,0 +1,1 @@
+//# sourceMappingURL=Navigation.js.map