+++ fiddles/angular2/fiddle-0021-SplitterComponent/src/component/Header.ts
@@ -1,0 +1,19 @@
+
+import {Component} from 'angular2/core';
+
+@Component({
+    selector: 'header',
+    template: `
+    <nav class="navbar navbar-default">
+        <div class="navbar-header">
+            <div class="panel-body">
+                <code>HEADER</code>
+            </div>
+        </div>
+    </nav>`
+})
+export class Header {
+
+
+
+}