+++ fiddles/angular2/fiddle-0021-SplitterComponent/src/component/Navigation.ts
@@ -1,0 +1,32 @@
+import {Component} from 'angular2/core';
+
+
+
+@Component({
+    selector: 'navigation',
+    template: `
+    <div class="col-md-2 r2d2-nav">
+        <div id="r2d2-navigation-el" class="panel panel-default" style="height: {{height}}px;">
+            <div class="panel-body">
+                <code>LEFT NAV</code>
+                <img src="resources/images/x.png" width="100%" />
+                <img src="resources/images/x.png" width="100%" />
+                <img src="resources/images/x.png" width="100%" />
+            </div>
+        </div>
+    </div>`
+})
+export class Navigation {
+
+    constructor() {
+        this._height = window.innerHeight - 50;
+    }
+    get height() {
+        return this._height;
+    }
+
+    set height(val) {
+        this._height = val;
+    }
+
+}